refactor(admin): split src/lib/admin/api.ts mirroring server.ts split (CHAOS-1571)
